ECS work so well omfg

+++ include/Components/PhysicalBody.hpp
@@ -20,15 +20,15 @@ namespace Components {
 			RIGHT = 270
 		};
 	public:
-		PhysicalBody(unsigned int xs, unsigned int ys, unsigned int zs) : x(xs), y(ys), z(zs), direction(DOWN) {};
+		PhysicalBody(float xs, float ys, float zs) : x(xs), y(ys), z(zs), direction(DOWN) {};
 		~PhysicalBody() override = default;
 		void summarize() const override
 			{ std::cout << "PhysicalBody | X : " << x << " | Y : " << y << " | Z : " << z << " | Direction : " << direction << std::endl; };
 		typeComponent getType() const override { return PHYSICALBODY; };
 	public:
-		uint32_t x;
-		uint32_t y;
-		uint32_t z;
+		float x;
+		float y;
+		float z;
 		Direction direction;
 	};
 }
